[testsharder] Add summary_if_skipped field in output
Create a summary_if_skipped field to replace the summary field. This is
done because determining if a shard is skippable is now done by looking
at the skippable field instead of checking if the summary field was
populated.

The summary_if_skipped field will be used to upload results to resultdb
of additional shards that the fuchsia.py recipe decides to skip
(additional to the list of shards that testsharder decides are
skippable) and will replace the summary field to be more explicit about
under what conditions this field is used.

Bug: b/450062147
Change-Id: I427568731e57dbd2eba4e236af734897ea439133
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1432754
Commit-Queue: Carver Forbes <carverforbes@google.com>
Fuchsia-Auto-Submit: Carver Forbes <carverforbes@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>