[audio][inspect] Make BufferTracker thread-safe
The BufferTracker class can be accessed from multiple threads,
but it was not thread-safe. This could lead to data races when
buffers are submitted and completed concurrently.

This change introduces a mutex to protect the internal state of
`BufferTracker`, ensuring that all access to its member variables is
properly synchronized. Additionally, `ZX_ASSERT_MSG` calls were
replaced with `FDF_LOG` to prevent crashing the driver in case of an
unexpected state.

Test: audio-inspect-recorder-test
Bug: 462554546
Change-Id: I3dd3f1b88e4487e038fd9c95d8c5000991ebf348
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1429571
Reviewed-by: Martin Puryear <mpuryear@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>