[lowpan] Fix input type for Verhoeff checksum calculation.
The otVerhoeffChecksumCalculate API takes aDecimalString as the first
parameter, which should be a string containing decimal digits converted
from the ephemeral key byte array.

Instead of passing a pointer to the first byte of the raw array, we need
pass a pointer to the string representation. This CL converts the key
array to a decimal string and passes the string pointer.

Bug: b/373664917
Test: manual test with Thread Utilities app to verify the overall ePSKc
      work flow, works as expected.
Change-Id: I6176d391ae9fda35d076180880bf51ad86ea73b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420636
Reviewed-by: Jiaming (Charlie) Wang <jiamingw@google.com>
Commit-Queue: Xiao Ma <xiaom@google.com>