[starnix] expose real block devices
When starnix is routed a /block directory, it takes anything in it and
puts it in /dev/block, using the semantic label it was routed as in the
block dir. Starnix handles converting file reads and writes to the
device into fuchsia block client reads and writes.

Bug: 450325631
Bug: 407091711
Change-Id: I113ec8712c7c1a99cd967f110526752e2ac939be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1425841
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>