[wlan][test] Don't write a trace if a test exits normally
Writing a traces for passing tests is a waste of resources. We can use
other means to generate traces for passing tests when we need them for
comparison with failing tests.

A trace will be written for a test if the test did not complete within
2 minutes, the streamed bytes to the trace file exceeds 50MB in size,
or the test panics.

Bug: 447408672
Test: Manually verified a trace is recorded in each scenario.
Change-Id: I50b44f0a11c5959ad8e10942d05839ccae1d766c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426754
Reviewed-by: John Tan <johntan@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>