[wlan][hw-sim] Create stress test to tease out flakes in hw-sim startup sequence
We've seen flakes in the startup sequence of a hw-sim test, specifically
while waiting for the WlanSoftmacStart event. This test runs the startup
sequence 50 times in an attempt to trigger a flake more often.

Inclusion of this test is temporary until we have more data on the flake.

Bug: 447408672
Change-Id: I17852730c039628248f771c87e322dd0f617035f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426757
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>