[auth] Ensure ADC in preflight
All supported auth mechanisms still require valid ADC,
so it should be covered by preflight check.

This should help with first-time setup cases.

In the event that the ADC exists but is corrupted or causing errors,
there is always `fx rbe auth -f` to force-regenerate it.

It might be possible that a previous invocation of bazel
has cached stale credentials, which can be cleared with `fx bazel
shutdown`.

Bug: 405454777
Change-Id: I5a4f9f4b1565b48022b39c3e152a9e53287bed43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433155
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>