[network/tests] Only parse real TCP segments
Update the socket integration test's TCP parsing logic so that it
only tries to parse TCP segments out of data that *is* a TCP segment.

The original logic would try to parse a TCP segment out of every packet,
and expect a failure if the packet didn't carry a TCP segment. That's
not guaranteed behavior and shouldn't be relied on.

Change-Id: I42a30ced3b6d8d91e6c722bdb78cc256a5fb79d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433854
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>