[fxfs] Ensure superblocks aren't confused with earlier filesystems.
If we find a valid-looking B superblock with a generation greater than
the A superblock, we will assume it's newer *and from the same
filesystem*. This can lead to problems if we overwrite one filesystem
with a new one using image_builder_mode, where we leave the second
superblock intact.

This CL tries to read the highest superblock generation. If they don't
exist, that's fine, we use generation 1. If they DO exist, we use
generation + 1 to ensure that the newly written superblock supersedes
all previous ones. Superblock generations are wrapping so this works
for all generation values.

Change-Id: I2f7c2ae575cff72f51badaf28a91cebbf7660c43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433239
Commit-Queue: Aaron Drew <ripper@google.com>
Reviewed-by: Chris Suter <csuter@google.com>