[memory_monitor2] Fix test broken by missing config
In change 1416614 we removed the dependency of memory_monitor2 on its
assembly-provided configuration (because it is no longer
configurable); as a result we removed the inspect node where we would
log the current configuration before.

We neglected to update an assertion in one of the end to end test that
verifies that the node in question is always populated.

Change-Id: I348cb141d7aaa741e67b0c404f34e6fef5f3bfdd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433894
Reviewed-by: Pierre Labatut <plabatut@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>