[wlan] Create core e2e test for SAR settings
This CL introduces a SAR settings test, which validates that:
1. all SAR settings can be set via the core APIs
2. the device connects to an AP
3. the SAR setting is persisted after the connection

This test is written in sync python. To facilitate that, a new sync copy
of the Core Testing base class is added. This class (and the test) make
extensive use of Honeydew to avoid code duplication.

Test: fx test sar_setting_test
Bug: 463663537
Change-Id: I0496c9a25d9210c196cfb34bfb715f55d01db45d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433095
Reviewed-by: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>
Commit-Queue: Marc Khouri <mnck@google.com>