[check-licenses] Update fmtlib license path
The license file for fmtlib dropped the "rst" file type several years
ago, so should update the path to remove the suffix as well.

For unclear reasons, https://fxrev.dev/1418918 verified this and failed
to build on some products due to this license not being present.

Change-Id: I00f0183053a97872fe9ae3b65135e2b8016579fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1434080
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Jerry Belton <jcecil@google.com>