[blobfs] Use blob protocols in tests
This change adds helper code for using the blob protocols in blobfs
tests. Not all of the helper code is used in this change but it will be
in later changes that switch over other tests.

Several of the easier integration tests have been switched over to use
the blob protocols in this change.

ExecutableMountTest has been removed because blobfs doesn't serve
executable VMOs over BlobReader. pkg-cache is responsible for making
VMOs executable with BlobReader.

Bug: b/337261333
Change-Id: I57edf81f23011a62493b4f0a1aec2effd86079b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1426974
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>