[sestarnix] Avoid iterating over type IDs
We had been iterating over the (sometimes long) extensible bitmaps
specifying types (and transforming the integers they represented, and
gathering them into a HashSet) just to make a membership query. We can
do the same membership query with a view over the underlying
ExtensibleBitmap.

(We could do the same for user IDs and role IDs, but we don't at this
time have any performance data indicating that that would measurably
improve any of our current executions.)

(We still iterate over and collect the integers in the error-handling
path, but that doesn't have performance implications.)

Fixes: 458317188
Change-Id: I3a189b6441c69af1095cd5008dbdbb46769c9443
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1428734
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Nathaniel Manista <nathaniel@google.com>