[components] the target of a route is required
This commit moves the target WeakInstanceToken from the Request struct
of a route to the argument set, and makes it required. This change means
that whenever a capability route is to be performed, the target (aka
initiator) of the route must always be supplied, and a default value
cannot be relied on. This is necessary because sometimes default values
for routes are set at a different point in the component graph than
where routes originate. For example if a dictionary is exposed and then
accessed through the exposed directory, when an entry in the dictionary
is opened the "default" target would be set where the capability was
offered into the dictionary, instead of where the exposed directory was
accessed.

Change-Id: I48e567ce5f34afa92694f3541f0b92c8c7c66a1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1430014
Reviewed-by: Olivia Nordquist <nolivia@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>