[mkbootimg] import to third_party
This repo contains some tooling that is useful to create and analyze
Android boot images.

We already have a copy of mkbootimg.py in the tree but it would be
better if we could switch to a more formally imported/versioned copy
here. Additionally this repo contains unpack_bootimg.py which is useful
to analyze and dump boot images.

Initial version is android16-qpr1-release-fuchsia, which is the same as
the most recent AOSP push android16-qpr1-release but with a cherry-pick
to include a LICENSE file.

Change-Id: I185db4d7f66c7a9c70e8bf7cb2f0bae258536e19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1430735
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>