[ktrace_provider] Quiet Logs About Exceeding Profile Allocation
Ktrace provider uses a thread profile to ensure it's able to timely copy
data. Under high memory pressure and/or cpu load, it can consistently
fail to finish its copy of the trace data within its allocation, causing
log spam.

This patch quiets the warning to debug. When viewing a trace, it's
obvious when we drop kernel data due to not being able to copy it fast
enough -- we get events emitted from the kernel showing the spans of
time we dropped records.

We could go a step further and keep a running average of how long it
takes to copy the data. However, because we have kernel trace data, we
can use that the compute in-post the time the ktrace_provider took to
copy the data, either by using the syscall events to measure
zx_ktrace_read calls, or scheduler events to see how long it was
scheduled for.

Bug: 465141469
Change-Id: I88c92f553bf243fb5ee64252636fb5e9c453c46c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1438015
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>