[ffx][doctor] Add check for ffx-usb-driver status
This change adds a new check to `ffx doctor` to report on the
status of the `ffx-usb-driver` process. It checks if the process
is running, reports its PID and the socket path it is listening on.

It also warns the user if the discovered socket path does not match the
path configured for ffx.

Test: Added unit tests. Manually invoked ffx doctor.
Bug: b/460195632
Fixed: b/460195632
Change-Id: Ia9cffafa5ad1943d8c86c589e4bc40c76a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1420157
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>