Reland "[botanist] Restart the ssh controlmaster when detecting a reboot."
This is a reland of commit d081217f14dd484650e63794ff477852227c06c8
This differs from the original CL by having testrunner call to restart
the ssh controlmaster before each test if it detects that it's not
running as opposed to botanist continuously restarting it. This will
allow us to reconnect to the target as well if we detect that the SSH
connection was lost whereas before we only try to reconnect after a
failed test.

Original change's description:
> [botanist] Restart the ssh controlmaster when detecting a reboot.
>
> Fixed: 463446410
> Change-Id: I9ef6ea1a8ba2d27061ba4752ae8c70c497403c85
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1433314
> Commit-Queue: Ina Huh <ihuh@google.com>
> Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
> Reviewed-by: Carver Forbes <carverforbes@google.com>

Change-Id: I8352b99bed12b92d432720ecf809f8023312c486
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1434301
Reviewed-by: Carver Forbes <carverforbes@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>