[dwmac] Improve shutdown if Create fails
If the call to Create, which is called when the driver binds, would fail
at certain points it would cause a crash. The dispatchers that were
created would not be properly shut down and then their destructors would
trigger an assert.

This moves the creation of dispatchers after the creation of the
deferred cleanup used to clean up if an error happens. This way the
dispatchers don't exist and therefore can't be destructed until there
is a proper cleanup handler.

This also fixes another issue where the Create call held on to a unique
pointer for too long after DdkAdd. If Create failed between DdkAdd and
the release of the unique pointer the object would be destroyed at the
end of Create and DdkUnbind would then crash because the object was no
longer valid.

Test: Manual testing by inserting failures in Create
Change-Id: Ib256a3fecfd8e5a4d2434a5e2562595d97551d7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1438300
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>