[starnix] Update android bootreason to kernel_panic when the session restarts
Android bootreason captures the boot reason of Android from the
underlying operating system. As of now, Fuchsia determines this based on
the RebootInfo.
However, though it isn't captured by the RebootInfo, an abrupt session
restart (typically caused by the Starnix kernel crash) should be
categorized as kernel_panic as well as it leads to Android restart. This
change is to implement this.

This change adds:
- Generate a file in the tmp directory when a new Starnix kernel
  component is created. With that, the later Starnix component can use
  it to identify whether or not itself is the first Starnix component
  running after the boot.
- The tmp directory to be used is taken from the container namespace.

Bug: 394953531
Test: android_bootreason is set as "reboot,cold" in the first session,
  and is updated to "kernel_panic" after a session stop-restart
  $ ffx session stop
  $ ffx session start
Signed-off-by: Dodam Kim <dodam@google.com>
Change-Id: Iab99b1d5895b04798339419461efb01c0cee46f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1412514
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>