[cog] add git polyfill for use in workspaces
This CL adds a git wrapper which will be used inside of the cog
workspace. This script does not add any commands and is not wired up to
fx in this CL. That logic will be added in a later CL. To test, you can
just execute the script directly.

Bug: 461899545
Change-Id: I1a92c9f6af7b8b344d3a26f3e3651f35e85ee5d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1438302
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>