[debugger] Ensure threads are synchronized on exception
Now that DebugAgent is correctly *not* attached to individual process
exception channels in certain modes, we are not receiving thread
starting events.

An implication of this is that if there is an exception in a process
underneath a job we're watching, we have to be very sure that we
actually have a DebuggedThread object for it, which means we might need
to synchronize the DebuggedProcess with the underlying Zircon source of
truth before collecting information that some FIDL clients might expect.

Change-Id: Ib29767100a66d57b1807f5cf438012b76a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1435995
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>