[fastboot] Allow "erase userdata" to change "flash blob" target
Allow "erase:userdata" to change the target of "flash:blob" to "super"
similar to what we do with "update-super:super:wipe". This allows using
either command with Android flashing scripts. Ensure that
"update-super:super:wipe" shreds the data volume as soon as the command
is run, instead of relying on a subsequent "flash:blob" command to
overwrite all of super + userdata.

Remove ability of subsequent "update-super:super" commands without the
":wipe" argument to restore the original target for "flash:blob" since
it isn't necessary (there's nothing to preserve if we already wiped
userdata).

Flashing the blob volume will still fail if the super partition was
never provisioned on a device, and we never wiped the userdata volume.
This will be addressed in a follow-up.

Bug: 458075061
Test: fx test fastboot
Change-Id: Ie14a361fb42124325ddc3167beebd73307c0ae9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1435397
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>