[debugger] Allow `continue` over exceptions
..instead of repeatedly hitting the same exception over and over again.

This helps exceptions in tests in particular which are effectively stuck
until a user does some form of a "detach" event.

In the CLI interface this is done by typing "quit" or "detach", but in
VSCode there's only one "detach" button, which is not obvious that that
is the one to press to get out of the debugging session. Now the "play"
button will cause the test to exit after two presses, due to
second-chance exceptions. Handling those is done in a followup, as well
as enabling the other buttons to not get stuck.

Bug: 444070494
Change-Id: I29c810c62b134537dacb2b81c586f4f66a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1421474
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>