[driver_runtime] Include threadid and role in stall message
Includes the threadid that was stalled and the role of the thread
pool when warning about a stalled thread to make it easier to
attribute the message to a cause.

Also lowered this message to an INFO level message since it's not
really a critical problem until all threads are stalled, and that's
still a warning.

Bug: 442604405
Change-Id: I53733d6d5c165c4275017e06c72ad8252a115682
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1435398
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>