[netlink] Avoid dropping NLM_F_DUMP responses
This is a simple workaround to address https://fxbug.dev/459883760.

When sending a Netlink message from starnix to the client, if the
clients receive buffer is not large enough to accommodate the message,
double the receive buffer size (up to 4MiB). This ensures correctness,
at the cost of additional memory used by certain netlink sockets.

Bug: 459883760
Change-Id: Ib4105a1333178f7c00e2cedf736ce7b35eb4d036
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1438115
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Sergey Ulanov <sergeyu@google.com>