[fxfs] Clean up dirty bytes when dropping unflushed files.
When files are modified after unlink, dropping the last reference
purposely doesn't flush the file since the contents are about to be
purged. This is the case for dropping unnamed temporary files as well.

When pages are dirtied past the end of file they cannot be safely
dropped since there may be an appending write that touches those pages
between. So these dirty ranges are tracked through the file lifetime and
dropped during close. Previously those bytes were leaked in the
accounting in the VolumesDirectory, now they're reported back to the
VolumesDirectory to stop tracking them when they get dropped.

Adds in a debug assert to check that there are no pages left that need
flushing when the vmo is being dropped, and any dirty pages that don't
need to be flush, are marked as clean in the accounting since they're
going away.

Bug: 452935329
Change-Id: I381f17fd87489c773f81820ef098570f31ce6084
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1434874
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>