[debugger] Skip second-chance exceptions in tests
This makes use of the test_failure_stack_matcher to inform the process
object that it is a test, which lets its threads query the process for
additional help when deciding how to handle an exception. This creates
an illusion for tests that disregard second-chance handling and
correctly deal with debugger installed breakpoints, hardcoded
breakpoints, and other types of exceptions without needing information
about the specific test run time.

Importantly, this has no effect until we actually realize that we're a
test. This is important, because we don't want to skip second-chance
exceptions in all cases.

As long as we have a stack matcher for a particular runtime's test
assertion failure path, we can identify that we are a test.

Note that this does mean that the test process will crash "for real",
since the exception will be forwarded, rather than handled.

Bug: 444070494
Change-Id: Ib38ad02ec7267553de535bb97ecc7b796a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1409614
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>