[fidl][rust] Don't require enable_rust_drivers for 'next' bindings
This changes it so that driver bindings are always generated for
the rust next bindings if contains_drivers, allowing for generating
them in the wire bindings without also forcing the partial
support in the old bindings to be generated, which may cause errors.

From the driver framework team's perspective, the driver transport
support in the rust next bindings are as supported as the rust
next bindings themselves are.

Change-Id: Iffb7c804554eb885bde6c1657e88e0c392cb2a88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1439494
Reviewed-by: David Koloski <dkoloski@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>