[toolbox] Expose tool protocols to toolbox
Gather all the protocols currently routed from into the SSH namespace
and make them available to the toolbox instead, so we can have a single
namespace that concentrates the tools available for shells.

As we're doing this, adopt the toolbox pattern of moving all the offers
to the toolbox dictionary closer to where the components are inserted
in the realm.

Bug: 416063207
Bug: 447668986
Change-Id: I3a86bba2b15165569aa8ae93c3b5fcdd6a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1427114
Reviewed-by: Eric Stone <ecstone@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Owners-Override: Adam Barth <abarth@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>