[wlan] Add factory_address to fullmac FIDL query
This change adds a `factory_addr` field to the return value of the Query
method of WlanFullmacImpl protocol. This represents the factory MAC
address, as opposed to the mutable `sta_addr`. This is a pre-requisite
to enable MAC randomization.

This change also populates that field for the brcmfmac driver, returning
the `sta_addr` as the `factory_addr`, since MAC changes are not
currently supported for the brcmf devices.

Lastly, this change depends on the CL below, which stubs out that return
in the synadhd driver, in //v/g/.

Bug: 455612638
Depends-on: turquoise-internal:I1e72cad890a1c9202e9880aa81f6e0cc51160892

Change-Id: Ia0e76710239c8712b0ecd77fdaf8cb1bb1a67104
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1405994
Commit-Queue: Hayden Nix <haydennix@google.com>
Reviewed-by: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
API-Review: Marc Khouri <mnck@google.com>
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>