[fdomain] Convert Handle to arbitrary borrowed types.
With this change it's possible to run `as_unowned::<T>` for any
`T: HandleBased` for a borrowed usage of a handle as an abirary type.

This is primarily used in future fuchsia-controller changes where
FDomain handles are stored in a map for a given fuchsia-controller
`Context` object.

Bug: 465842888
Change-Id: I878f4cf53d370a3adff5fed617629b5d6a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440484
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>