[debugger] Step should not get stuck on exceptions
...which provides a much nicer UX flow when stuck on an exception, both
on the CLI and in VSCode.

Prior behavior continually tried to single step over an instruction that
raised an exception, which would get marked as handled and then step
again, raising the same exception, causing no progress to ever be made.

Now, we make use of second-chance exceptions to allow us to forward,
rather than handle, the exception, which still allows the user to take
action at the second chance, and for optimizations to be made in certain
contexts (like tests) to not get stuck on a test failure line.

Fixed: 444070494
Change-Id: Ie7efde7dee83a0bd5c163029ca912b716a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1411755
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>