[components] add OpenExposedDir to fuchsia.component.Controller
When a child component is created a fuchsia.component.Controller may be
attached to it. This allows the parent to interact with this child
without needing to constantly reference it by name, which somewhat
simplifies things and removes one possible error category.

Add OpenExposedDir to this controller protocol, to complement the
already existing GetExposedDictionary and mirror the functions on
fuchisa.component.Realm.

Change-Id: I8ce0f6f33e09337197acef9cd987e0d6028e946a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1432594
Reviewed-by: Olivia Nordquist <nolivia@google.com>
API-Review: Wez <wez@google.com>
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>