[socket-proxy] Fully clean up state on connection
A reconnection to the StarnixNetworks or FuchsiaNetworks protocol
indicates a reset of networking state for the socketproxy. Previously,
we were clearing the networks list, but not resetting the default
network id for the respective registry. This led to sockets attempting
to get the mark of the default network, but not finding any network in
the network registry matching the id.

Fixed: 465164164
Change-Id: I7213b4ad5b1be83c097f8e7ad50663fa696e9f10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441413
Reviewed-by: Zoey Bush <zmbush@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>