[kernel][vm] Use prev_eviction_result for same page check
For IncorrectPage eviction failure, the page will either be moved out of
the Isolate list when the ownership change completes, or it will remain
in place when ownership changes due to clone creation/deletion. In the
latter case, eviction is going to be attempted on the page again, which
is not a bug. To correctly account for this, use the previous eviction
result for the same page check, and update the previous result even if
the failure reason is IncorrectPage.

Bug: 434361683
Run-All-Tests: true
Change-Id: Ia346f7f0f736bd6b61280b58758a65c4e7043cb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441521
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>