[libc] Have MemorySnapshotStartArgOnSuspendedThread check all regions
This particular test checks that the start arg passed to a new thread is
accessible in the TLS region via __sanitizer_memory_snapshot. As libc
changes, the actual location of the allocation may change. Update this
test just to assert that the start arg is searchable via any of the
callbacks since it should be kept live.

Bug: 342469121
Change-Id: I54bd2f5ea978c9535cbe33286409334cbf74ab2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441748
Reviewed-by: Paul Kirth <paulkirth@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>