[dwmac] Increase reset timeout
Some devices seem to time out when querying the bus mode after reset.
Increase the timeout in an attempt to avoid this. For most devices this
should have no impact and for devices that are on the edge it should
should increase make them succeed every time.

For devices with faulty hardware or other issues this will slow down the
time to failure. Hopefully that will not impact a significant number of
devices.

Change-Id: I5f6bfcded21b49347dad7b53414b630a97bc0338
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441274
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>