[fuchsia-controller][honeydew] tracing refactor.
This refactors the tracing affordance to use fuchsia-controller more
idiomatically. This avoids the use of threads (which is likely to cause
undefined behavior with fuchsia-controller code, as it assumes a
single Python thread).

This also uses the async adapter code to clean some lines up that were
previously using `asyncio.run()`.

This also updates some of the fuchsia-controller documentation to
clarify some of the potential pitfalls in the existing libraries.

Change-Id: I669e129904038b909951952326419c276a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1440728
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>
Reviewed-by: Sam Balana <sbalana@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>