[log] Do not use a stack based buffer for sending log message
This buffer is responsible for 4Mo of permanent used memory in Starnix
when userspace logs to kmsg.
Moving this on the heap ensure this memory is reused using scudo buffer
instead.

This can probably be further improved by not initializing this memory to
0.

Bug: 466294903

Change-Id: I1b2d7066fde125be6a19792329c494174c17d067
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1441148
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>