[packet] Introduce SplitByteSliceBufferView
Expose an implementation of `BufferView<B>` on top of a SplitByteSlice.
This type will be used in a subsequent packet-formats change to parse
TcpOptions.

While here, fix a latent bug in the `SplitSliceBufferView::take_back`
implementation. Fortunately this code path was unused so there should
not be any behavior changes, but it removes a trap for future users.

Bug: 462085360
Change-Id: I08923c467bcf3ce84b83c4bc4c293cf90a591f53
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1442308
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>