[packet-formats] TCP Options Parsing/Serialization
Migrate the implementation of parsing and serializing of TCP Options off
of the "Records" framework. Benchmarks have revealed performance issues
with that framework, and a handrolled implementation of
parsing/serialization can be far more efficient when TCP options are in
use. This performance difference is particularly noticeable when a
connection uses the TCP Timestamp feature, as such connections have
options present in every segment.

From perfcompare, this change:
* Parsing with TCP Options (between a 1% & 5% improvement)
* Serialization with TCP Options (between a 0% & 17% improvement)
* Parsing without TCP Options (between a 1% improvement and a 2%
  regression)
* Serialization without TCP Options (between a 10% improvement and an 8%
  regression)

Bug: 462085360.

Change-Id: Ifb76f395e4b25854495851d936ca4fd53002cb78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1430375
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>