[netstack3] Improve route table `Debug` impl
The old `Debug` impl formats the pointer address of the table, which
is not very useful. This change improves the impl to output the
bindings ID, which is a numeric value.

Also change the bindings_id of a Core RoutingTableId to use an opaque
ID type instead of `u32`. This way Core has no obligation to verify
the given Bindings ID is unique and it is up to the Bindings to ensure
this.

Fixed: 462746973
Change-Id: I608f41648354f9820c3a12155085bff66a6a6964
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1438304
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>